+++ lib/Elastica/Aggregation/GeoCentroid.php
@@ -12,6 +12,16 @@ use Elastica\Exception\InvalidException;
 class GeoCentroid extends AbstractAggregation
 {
     /**
+     * @param string $name  the name of this aggregation
+     * @param string $field the field on which to perform this aggregation
+     */
+    public function __construct($name, $field)
+    {
+        parent::__construct($name);
+        $this->setField($field);
+    }
+
+    /**
      * Set the field for this aggregation.
      *
      * @param string $field the name of the document field on which to perform this aggregation